backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-jose[cryptography]
python-multipart
paho-mqtt
